requirements.txt
requests>=2.31.0
openai>=1.12.0
pyairtable>=2.2.0
tqdm>=4.65.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
supabase>=2.0.0
pinecone>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
starlette>=0.36.0

# Add Slack packages
slack-bolt>=1.18.0
slack-sdk>=3.27.0
